Conflicts:
	src/cpacs_other/CCPACSMaterial.cpp
	src/cpacs_other/CCPACSMaterial.h
	src/cpacs_other/TixiSaveExt.cpp
	src/cpacs_other/TixiSaveExt.h
	src/exports/CTiglExportIges.cpp
	src/geometry/CCPACSFarField.cpp
	src/geometry/CTiglAbstractGeometricComponent.cpp
	src/guide_curves/CCPACSGuideCurve.cpp
	src/guide_curves/CCPACSGuideCurve.h
	src/guide_curves/CCPACSGuideCurveProfile.cpp
	src/guide_curves/CCPACSGuideCurveProfile.h
	src/guide_curves/CCPACSGuideCurveProfiles.cpp
	src/guide_curves/CCPACSGuideCurves.cpp
	src/wing/CCPACSPositionings.h
	src/wing/CCPACSWing.cpp
	src/wing/CCPACSWingCSStructure.cpp
	src/wing/CCPACSWingCSStructure.h
	src/wing/CCPACSWingCell.cpp
	src/wing/CCPACSWingCell.h
	src/wing/CCPACSWingCells.cpp
	src/wing/CCPACSWingCells.h
	src/wing/CCPACSWingComponentSegment.cpp
	src/wing/CCPACSWingComponentSegment.h
	src/wing/CCPACSWingComponentSegments.cpp
	src/wing/CCPACSWingConnection.cpp
	src/wing/CCPACSWingConnection.h
	src/wing/CCPACSWingPositioning.cpp
	src/wing/CCPACSWingPositionings.cpp
	src/wing/CCPACSWingProfile.cpp
	src/wing/CCPACSWingProfile.h
	src/wing/CCPACSWingProfileCST.cpp
	src/wing/CCPACSWingProfileFactory.cpp
	src/wing/CCPACSWingProfileFactory.h
	src/wing/CCPACSWingProfilePointList.cpp
	src/wing/CCPACSWingProfiles.cpp
	src/wing/CCPACSWingSection.cpp
	src/wing/CCPACSWingSectionElement.cpp
	src/wing/CCPACSWingSectionElements.cpp
	src/wing/CCPACSWingSectionElements.h
	src/wing/CCPACSWingSections.cpp
	src/wing/CCPACSWingSections.h
	src/wing/CCPACSWingSegment.cpp
	src/wing/CCPACSWingSegment.h
	src/wing/CCPACSWingSegments.cpp
	src/wing/CCPACSWingSegments.h
	src/wing/CCPACSWingShell.cpp
	src/wing/CCPACSWingShell.h
	src/wing/CCPACSWings.cpp

+++ src/geometry/CTiglShapeCache.cpp
@@ -52,7 +52,7 @@ bool CTiglShapeCache::HasShape(const std::string& id)
 
 unsigned int CTiglShapeCache::GetNShape() const 
 {
-    return shapeContainer.size();
+    return static_cast<unsigned int>(shapeContainer.size());
 }
